+++ src/traits/region.rs
@@ -1,19 +1,19 @@
 use nalgebra::Point2;
 
 pub trait Region {
-    fn left(&self) -> i64;
-    fn top(&self) -> i64;
+    fn left(&self) -> i32;
+    fn top(&self) -> i32;
     fn width(&self) -> u32;
     fn height(&self) -> u32;
 
     #[inline]
-    fn right(&self) -> i64 {
-        self.left() + (self.width() - 1) as i64
+    fn right(&self) -> i32 {
+        self.left() + (self.width() - 1) as i32
     }
 
     #[inline]
-    fn bottom(&self) -> i64 {
-        self.top() + (self.height() - 1) as i64
+    fn bottom(&self) -> i32 {
+        self.top() + (self.height() - 1) as i32
     }
 
     #[inline]
@@ -22,20 +22,20 @@ pub trait Region {
     }
 
     #[inline]
-    fn contains(&self, x: i64, y: i64) -> bool {
+    fn contains(&self, x: i32, y: i32) -> bool {
         self.left() <= x && self.top() <= y && self.right() >= x && self.bottom() >= y
     }
 
     #[inline]
-    fn center(&self) -> Point2<i64> {
+    fn center(&self) -> Point2<i32> {
         Point2::new(
-            self.left() + (self.width() / 2 + 1) as i64,
-            self.top() + (self.height() / 2 + 1) as i64,
+            self.left() + (self.width() / 2 + 1) as i32,
+            self.top() + (self.height() / 2 + 1) as i32,
         )
     }
 
     #[inline]
-    fn top_left(&self) -> Point2<i64> {
+    fn top_left(&self) -> Point2<i32> {
         Point2::new(self.left(), self.top())
     }
